Addressed review comments from Joe S
