--- cert.pem
@@ -1,29 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFBDCCAuygAwIBAgIJAJR2Z5lKdQbaMA0GCSqGSIb3DQEBCwUAMBcxFTATBgNV
-BAMMDFJlZGZpc2hFdmVudDAeFw0xNzA1MjQxMjM5MDJaFw0yMjA1MjQxMjM5MDJa
-MBcxFTATBgNVBAMMDFJlZGZpc2hFdmVudDCCAiIwDQYJKoZIhvcNAQEBBQADggIP
-ADCCAgoCggIBAMNMB3v/ZUCNxgLUyw6jYKZLkyqpUYXAw4vw6s9+23PqgKnuc9oZ
-FG5GYheBp/lsXcPUXq/jDe4GE3gaZEWrVUjNiHH6EPbEd0WbMEqguxWFuXaECCSh
-k2PHI1EYhDWgm2IUgrrMzt5m0M32CMo3IjteuGHbAhfcvYKrSV0enMXWC0M3Mja9
-0GuMqp4JsyWpou/f+J7oXm8bV0uLUBNikAJlVqFAsg2diHh3e45y15CqGB5rJcGV
-MiVQSQ7LYKBryWs6+L0WoCO5N9pTm5fTJu5D+JWrdveDj50ZfJSXxR5bO9zP97uC
-HZfZshRjWnk+TcAxj1jhOgxe2AYWjWpxpjxu23hFmeu29DghmoFvWPjliJkPa3a2
-Y51hE3g1Skd9zMh5UNTWdbIn87XElpNBgpnVyim3EvNXcMeZ3vLXEW1lpw8enIiX
-eCDnJBCBViV24bTOk+gMybEs4Zp4kuT38epz5rp3jMqaIp383opNLBtQiZ5ex3MT
-bU6oBGjN6sSaW1zZf0amoQYw0c1lVbZ6U7Na4Eyrr+LSeyZw1GpNZHwV9xGEnVkP
-Psgdt3nXICnjasJ5CHX04iCsJwuvpDXrR3FwJiZYYm3LANkkspDoryrb/kQRlKTs
-kk4nTBk/i7yqWom2rzCXejg/Gbb6DyKUEPLUlFqI7O3Nvq92IxzXGJ0LAgMBAAGj
-UzBRMB0GA1UdDgQWBBTBB0Tyst7jJx5dvYF2wIOLB0nYiTAfBgNVHSMEGDAWgBTB
-B0Tyst7jJx5dvYF2wIOLB0nYiTAPBgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEB
-CwUAA4ICAQCgVYFHsCMmhHEur+38V/ciUokgL0TGxhkxZLmT+WdmyRC/I+6mM5mC
-yquh3qx8bVmDx24bD7VjkUdf+kXDrQkq1EfkSmXnCn8+eiShEPfzBAiU2gd0F6S+
-iuz3SJMhTseKMhyBwDRzRUraSqMCmYwaeG/nbRvYjRxMmz0zzxOAnwQhE5WFh0Yv
-vhbtKHxnoIjL/EZ84nS7tiDv41zd1se5l0effc2B0sE4PAU1dxWSTgMTFNybYHxb
-08YvSwXm2a4YHtIUw0hGf2CLB0TQWilXpksWs5N0p287n/4rOs03lvIjJZv1b0ZL
-B2GQ8C0PbIu7ZxPqePXPyDpsd2qv7LgYFCz1RdfjYzRRPbMu/5+NbOl1p4OVNY4i
-Xqy1j/zGa0tDw7DgtadNRa07UACBHhRORLFdbLJAezaKraXrota6TGl88SoaIxPv
-ujxOoSpD09g5zV82YxsV27m7cnsqIwuvt2yG66qd8E0v1MR6ln7r8qEbifZ6qVhv
-p93LmxGDTmrLjVWXG7QFwffsIabCPQwHgRXSq796wm/HxTpJ+1VtxT8ABdo6N6Ur
-utG1ltmpeislDagY0McMy0gAXqHfBb318ZprR0UemuGy6G4C74PZBKQtGjCf82iV
-B6jpbdZdvti3oGFrYTqgZ3DcGfvgALm5dci7TQOPmeACbmIWoGvuuA==
------END CERTIFICATE-----
