+++ redfishAPI.cfg
@@ -18,9 +18,9 @@
     "urltimeout": "30",
     "usefinalslash": "True",
     "username": "",
-    "listenerusessl": "False",
+    "listenerusessl": "True",
     "listenerip": "localhost",
     "listenerport": "8080",
-    "certfile": "cert.pem",
-    "keyfile": "server.key"
+    "certfile": "",
+    "keyfile": ""
 }
