--- server.key
@@ -1,51 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIJKgIBAAKCAgEAw0wHe/9lQI3GAtTLDqNgpkuTKqlRhcDDi/Dqz37bc+qAqe5z
-2hkUbkZiF4Gn+Wxdw9Rer+MN7gYTeBpkRatVSM2IcfoQ9sR3RZswSqC7FYW5doQI
-JKGTY8cjURiENaCbYhSCuszO3mbQzfYIyjciO164YdsCF9y9gqtJXR6cxdYLQzcy
-Nr3Qa4yqngmzJami79/4nuhebxtXS4tQE2KQAmVWoUCyDZ2IeHd7jnLXkKoYHmsl
-wZUyJVBJDstgoGvJazr4vRagI7k32lObl9Mm7kP4lat294OPnRl8lJfFHls73M/3
-u4Idl9myFGNaeT5NwDGPWOE6DF7YBhaNanGmPG7beEWZ67b0OCGagW9Y+OWImQ9r
-drZjnWETeDVKR33MyHlQ1NZ1sifztcSWk0GCmdXKKbcS81dwx5ne8tcRbWWnDx6c
-iJd4IOckEIFWJXbhtM6T6AzJsSzhmniS5Pfx6nPmuneMypoinfzeik0sG1CJnl7H
-cxNtTqgEaM3qxJpbXNl/RqahBjDRzWVVtnpTs1rgTKuv4tJ7JnDUak1kfBX3EYSd
-WQ8+yB23edcgKeNqwnkIdfTiIKwnC6+kNetHcXAmJlhibcsA2SSykOivKtv+RBGU
-pOySTidMGT+LvKpaibavMJd6OD8ZtvoPIpQQ8tSUWojs7c2+r3YjHNcYnQsCAwEA
-AQKCAgEArYT+3WBFIT8WIOvaYu4scHSTlIMOKM3lsI/kwWmQKITBvrpb+vrvmS81
-9na/vTBYeydU3efaN5L0d3DTiMc7Jk1fnw01R1u2MYU6OgRnVS8UYsXaylk5MJz9
-tw8WYOOLS3C32qsgXqBxg53Hn4uvAZseLeu+B1+cjd4MeA0789dC+HyxQ8wQw44f
-Bf3OS3dWitSv/NivFv7u5/MmuaKIG2jYuLtXuBsw5wAaKq3E9Ab9GQHQcRhTjq2v
-y9JNzxsVwg2Kdc8Fh9gIVRaWFyaH34lK9Ah90lr1HftIImpavKcJIB8zmEl6a7ID
-tHO2LiV6RXqDR9ANmHIoc/hTu7CwplHlEIvzc/oXs1ac6Dabtwea1aHzZPYt8TBi
-NsuA0AUlmbuE1m8UUqG+6bFtomTcTtbVkSkspoT6jcfPDbQg7pmfXFmoJXqTaBWj
-Bd+ovL3sCCFsA0MRDg22oreDeXvqNgq7oct0ABdwvnPBTbakCZ/3PoShWn2kJevG
-gfTxgmYqIaqkiork6HVzpqq4KZqH2RZfAl7CLuu8gd3oLS3c2dMgHdrrM4qwFEPS
-WcFTUoZJlxSZ0M2o3xFjTcqQgcKkPuu0BO8E0Ph1jrhRUJMFPT+1tJjExFJfgEdy
-uaenf/Cllm3Q9ciso4yK3Va3QV3ocY+6WlOl3/gUxU0CuNblLskCggEBAOLDcKs3
-nQbkVBFOf4kv+AqrvrwrxZQfP6LvpPoCFmcRtVpl4eXN49LzR5hDIbtqhCrsdeug
-LWf/7ZWMqQvNSUHRPtsHSPHK7JR/KZRa9rTLQ/VO5Gb5XO1+U/S1TfHzswUZMF7t
-F3PfUtWg7Ywjr5xwh4hh8Ph+kh2tNZUEqCv8G9kptF1bPu/dqibpBtqlrNaRYtBM
-pZ8cM3l3DzmDv4JD287f0ipXsju9QdYBu6WIkEknDb5zWbyzh7UsaepGLJhWstvU
-WFrHX6M7ZmPyucSlnmTfXCLipVI6z/csHp0d3jwBCktksff5q0ZjcmPmQ6P7Xugc
-M6DrnXrVpnDDS50CggEBANx6AbdWxF1d1nfv+AScNYgXdM283Ps7RdH/MpiLmHic
-ObZPloXhXqmMEEgJBXmPp+Ipn9cnCGJ5JPT4Rf2Dvt68kt9CMen3hxFkTHDXPFKZ
-do/keE1njF+ekpqBolVu/VstuyoXfdB6+HIDDLxh73OSdZ7rIOeKMkOZoILhSWLY
-wSiLC2E0bI7fBoUy+P0IHU6WebR3zXcznFvsk/sEEnhfPyzf6xAhTaOE2ni9ujEd
-ptz5FaHerSUNCcsStU8Sij/vj1Q5JQkktICFJbCS4q6edzGF2K2aPvmkNdOqQef7
-0nTLz0jZ0MiN7SPBUX4cqXuFokwJuoyKWN3HyiCLTscCggEAcpgvtOba3OPANCrG
-Oq7aG5JlRjScmVkdUJrodUtexOFnE3LPMNHWYwF8Wc77GfofJo3unY+JwrfL+bpm
-p0cr0S4MJKXO0BoranPEsGszp3PdOPeQZxV3R7GCffStmir8egQbtKz+IBlog/ra
-cOqL0XPwmhRcsUdNUd/5ZwA+eFPkQt9XfaIh8RfE5/r08sOsz5YTKI7DV9/ph2WK
-dRdGTW87gpULYUiGErh2r86aZHHZ8I3i3S7v13OmjgEO2UxzNt0BcCQkLmP4ycW3
-3aSJCn5jyDg0GJEO2LpN5GwQYHf6mbow2J7fbSBn1J9m7GMuJY1fjc7KrHTFVZDo
-lq2q5QKCAQEAtYvtj9dtuGgFqxs5gipS1i5iautCl8k18/aM5hnEwkP6LcoZa0Kp
-3TjNS1PKY0/McSnPSa4FdHjG2sT5dQigCe8Djb2HNl4xpTuOyjghnu5pN5FtEmfa
-qf38QCfQ0UKkcwkUHlPbsOkhXVhvi0uljegi3GpG3QHp48pD5qG1+SP84fKkuzUc
-0wSI7PONHWYqI8U6kEU6NhaUgQkszTyMLPxZHA2CFPUnZieGcnfgMoN1HG4qxvCC
-aIj08HgGPgdJOd58imDHw2CrFU0OPPimmh04jtnlEwZ5Jr116UZAINM2eCA/HoDl
-L8/WXmPYQ908Hc9rVeuI5iMhi/Xw1Z2AjwKCAQEApIoJD2j/bkqPtn5MMT1TDgH2
-+ZFh0cpfTM0rKU1TiOL0+2MdLd1VcpdwSaDVp8O7a0zQ5vs3NMxiUAl5aS2NTzjA
-mYiYI+2wu21lu+QIpjwGJuKF7qKuAtw7GhUMMfEX/wlF5mI0oViWumBOtGxLJCH2
-vWsCd3Ft7cQP2O/h7PFRNZKobFpBDjf6s2ArIOP6qcQh0HSPB6pSW7eX2RUYWOnY
-QdL6PRdjKTBTc0hKuOy4R7Z0+WNfW476LzTZD31R71lRF1hGaJOaUtH0LgtPEDMo
-2hVmRSYbvM1IxZe/tHYpu3HH+GWDvN+SIN0n31tcyhskPnnIjF0C386psJYpbw==
------END RSA PRIVATE KEY-----
