+++ version.py
@@ -13,4 +13,4 @@
 # ******************************************************************************************
 #
 
-__version__ = '2.4.2'
+__version__ = '2.5.0'
